+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/blob_options.hpp
@@ -1373,12 +1373,68 @@ namespace Azure { namespace Storage { namespace Blobs {
     Azure::Core::Context Context;
 
     /**
+     * @brief Optionally specifies the offset of range over which to list ranges. This offset must
+     * be a modulus of 512.
+     */
+    Azure::Core::Nullable<int64_t> Offset;
+
+    /**
+     * @brief Optionally specifies the length of range over which to list ranges. The length must be
+     * a modulus of 512.
+     */
+    Azure::Core::Nullable<int64_t> Length;
+
+    /**
+     * @brief Optional conditions that must be met to perform this operation.
+     */
+    BlobAccessConditions AccessConditions;
+  };
+
+  /**
+   * @brief Optional parameters for PageBlobClient::GetPageRangesDiff.
+   */
+  struct GetPageBlobPageRangesDiffOptions
+  {
+    /**
+     * @brief Context for cancelling long running operations.
+     */
+    Azure::Core::Context Context;
+
+    /**
      * @brief Specifies that the response will contain only pages that were changed between
      * target blob and previous snapshot. Changed pages include both updated and cleared pages.
      * The target blob may be a snapshot, as long as the snapshot specified by PreviousSnapshot is
      * the older of the two.
      */
     Azure::Core::Nullable<std::string> PreviousSnapshot;
+
+    /**
+     * @brief Optionally specifies the offset of range over which to list ranges. This offset must
+     * be a modulus of 512.
+     */
+    Azure::Core::Nullable<int64_t> Offset;
+
+    /**
+     * @brief Optionally specifies the length of range over which to list ranges. The length must be
+     * a modulus of 512.
+     */
+    Azure::Core::Nullable<int64_t> Length;
+
+    /**
+     * @brief Optional conditions that must be met to perform this operation.
+     */
+    BlobAccessConditions AccessConditions;
+  };
+
+  /**
+   * @brief Optional parameters for PageBlobClient::GetManagedDiskPageRangesDiff.
+   */
+  struct GetManagedDiskPageRangesDiffOptions
+  {
+    /**
+     * @brief Context for cancelling long running operations.
+     */
+    Azure::Core::Context Context;
 
     /**
      * @brief This parameter only works with managed disk storage accounts. Specifies that