+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/page_blob_client.hpp
@@ -215,6 +215,28 @@ namespace Azure { namespace Storage { namespace Blobs {
         const GetPageBlobPageRangesOptions& options = GetPageBlobPageRangesOptions()) const;
 
     /**
+     * @brief Returns the list of page ranges that differ between a previous snapshot and this page
+     * blob. Changes include both updated and cleared pages.
+     *
+     * @param options Optional parameters to execute this function.
+     * @return A GetPageBlobPageRangesResult describing the valid page ranges for this blob.
+     */
+    Azure::Core::Response<Models::GetPageBlobPageRangesResult> GetPageRangesDiff(
+        const GetPageBlobPageRangesDiffOptions& options = GetPageBlobPageRangesDiffOptions()) const;
+
+    /**
+     * @brief Returns the list of page ranges that differ between a previous snapshot url and this
+     * page blob. Changes include both updated and cleared pages. This API only works with managed
+     * disk storage accounts.
+     *
+     * @param options Optional parameters to execute this function.
+     * @return A GetPageBlobPageRangesResult describing the valid page ranges for this blob.
+     */
+    Azure::Core::Response<Models::GetPageBlobPageRangesResult> GetManagedDiskPageRangesDiff(
+        const GetManagedDiskPageRangesDiffOptions& options
+        = GetManagedDiskPageRangesDiffOptions()) const;
+
+    /**
      * @brief Starts copying a snapshot of the sourceUri page blob to this page blob. The snapshot
      * is copied such that only the differential changes between the previously copied snapshot
      * are transferred to the destination. The copied snapshots are complete copies of the original